add level 1-2
This included a major change to rule parsing.

Level 1-2 is the first level to include "and", which my existing rule
parser did not support.

Since my existing parser did not seem like it would scale well to the
full grammar of Baba Is You, I decided to try implementing it in a
different way. I ended up spending a long time exploring a few different
approaches that either didn't work or were unsatisfactory. My final
solution was actually fairly similar to what I started with, but looks
cleaner now.

Along with adding support for "and", I added support for cells that
contain multiple pieces of text. Handling these is a tad complicated,
because as far as I can tell you just need to try parsing all possible
permutations of them. But this is implemented now. (Well, it's untested,
so could be wrong. But we'll deal with that when we need to.)

Finally, I also made a change to the text fomat for levels. Each level
now has a legend for all of the objects in that level. I decided to make
that change when I had to add three new objects for level 1-2 and found
that I was running out of characters to abbreviate them all. There is no
longer a global mapping from character to object, so collisions between
object names is less of a problem.

This change also added the first tests for this codebase. This means
that `cargo test` now does something useful.