fix goldens
in a previous commit

  (a963c500c038b829b38eaf9c5a8cca633350ecdc add level 5-3)

I had fixed the behavior of shift as follows: when eg a belt tries to
shift something eg keke and fails because the neighbor is blocked by
something Stop, the belt still changes keke's direction.

There were two golden files that recorded the old behavior in which the
direction was unchanged. Those files are now fixed.